+++ test/opf.jl
@@ -5,13 +5,7 @@
         result = run_ac_opf("../test/data/case3.m", ipopt_solver)
 
         @test result["status"] == :LocalOptimal
-        @test isapprox(result["objective"], 5812; atol = 1e0)
-    end
-    @testset "3-bus case with active DC Line" begin
-        result = run_ac_opf("../test/data/case3_dc.m", ipopt_solver)
-
-        @test result["status"] == :LocalOptimal
-        @test isapprox(result["objective"], 5906.88; atol = 1e0)
+        @test isapprox(result["objective"], 5907; atol = 1e0)
     end
     @testset "5-bus asymmetric case" begin
         result = run_ac_opf("../test/data/case5_asym.m", ipopt_solver)
@@ -19,11 +13,11 @@
         @test result["status"] == :LocalOptimal
         @test isapprox(result["objective"], 17551; atol = 1e0)
     end
-    @testset "5-bus with active DC line" begin
+    @testset "5-bus with dcline costs" begin
         result = run_ac_opf("../test/data/case5_dc.m", ipopt_solver)
 
         @test result["status"] == :LocalOptimal
-        @test isapprox(result["objective"], 17756.17; atol = 1e0)
+        @test isapprox(result["objective"], 17760.2; atol = 1e0)
     end
     @testset "6-bus case" begin
         result = run_ac_opf("../test/data/case6.m", ipopt_solver)
@@ -47,7 +41,7 @@ end
         result = run_opf("../test/data/case3.m", ACTPowerModel, ipopt_solver)
 
         @test result["status"] == :LocalOptimal
-        @test isapprox(result["objective"], 5812; atol = 1e0)
+        @test isapprox(result["objective"], 5907; atol = 1e0)
     end
     @testset "5-bus asymmetric case" begin
         result = run_opf("../test/data/case5_asym.m", ACTPowerModel, ipopt_solver)
@@ -77,7 +71,7 @@ end
         result = run_dc_opf("../test/data/case3.m", ipopt_solver)
 
         @test result["status"] == :LocalOptimal
-        @test isapprox(result["objective"], 5695; atol = 1e0)
+        @test isapprox(result["objective"], 5782; atol = 1e0)
     end
     @testset "5-bus asymmetric case" begin
         result = run_dc_opf("../test/data/case5_asym.m", ipopt_solver)
@@ -108,7 +102,7 @@ end
         result = run_opf("../test/data/case3.m", SOCWRPowerModel, ipopt_solver)
 
         @test result["status"] == :LocalOptimal
-        @test isapprox(result["objective"], 5735.9; atol = 1e0)
+        @test isapprox(result["objective"], 5746.7; atol = 1e0)
     end
     @testset "5-bus asymmetric case" begin
         result = run_opf("../test/data/case5_asym.m", SOCWRPowerModel, ipopt_solver)
@@ -136,7 +130,7 @@ end
         result = run_opf("../test/data/case3.m", QCWRPowerModel, ipopt_solver)
 
         @test result["status"] == :LocalOptimal
-        @test isapprox(result["objective"], 5742.0; atol = 1e0)
+        @test isapprox(result["objective"], 5780; atol = 1e0)
     end
     @testset "5-bus asymmetric case" begin
         result = run_opf("../test/data/case5_asym.m", QCWRPowerModel, ipopt_solver)
@@ -166,7 +160,7 @@ end
         result = run_opf("../test/data/case3.m", SDPWRMPowerModel, scs_solver)
 
         @test result["status"] == :Optimal
-        @test isapprox(result["objective"], 5788.7; atol = 1e0)
+        @test isapprox(result["objective"], 5851.3; atol = 1e0)
     end
     # TODO see if convergence time can be improved
     #@testset "5-bus asymmetric case" begin