+++ test/pf.jl
@@ -1,24 +1,6 @@
 @testset "test ac pf" begin
     @testset "3-bus case" begin
-        result = run_ac_pf("../test/data/case3.m", ipopt_solver)
-
-        @test result["status"] == :LocalOptimal
-        @test isapprox(result["objective"], 0; atol = 1e-2)
-
-        @test isapprox(result["solution"]["gen"]["1"]["pg"], 148.0; atol = 1e-1)
-        @test isapprox(result["solution"]["gen"]["1"]["qg"], 54.6; atol = 1e-1)
-
-        @test isapprox(result["solution"]["bus"]["1"]["vm"], 1.10000; atol = 1e-3)
-        @test isapprox(result["solution"]["bus"]["1"]["va"], 0.00000; atol = 1e-3)
-
-        @test isapprox(result["solution"]["bus"]["2"]["vm"], 0.92617; atol = 1e-3)
-        @test isapprox(result["solution"]["bus"]["2"]["va"], 7.25886; atol = 1e-3)
-
-        @test isapprox(result["solution"]["bus"]["3"]["vm"], 0.90000; atol = 1e-3)
-        @test isapprox(result["solution"]["bus"]["3"]["va"], -17.26711; atol = 1e-3)
-    end
-    @testset "3-bus case with active DC line" begin
-        result = run_ac_pf("../test/data/case3_dc.m", ipopt_solver, setting = Dict("output" => Dict("line_flows" => true)))
+        result = run_ac_pf("../test/data/case3.m", ipopt_solver, setting = Dict("output" => Dict("line_flows" => true)))
 
         @test result["status"] == :LocalOptimal
         @test isapprox(result["objective"], 0; atol = 1e-2)
@@ -31,9 +13,10 @@
         @test isapprox(result["solution"]["bus"]["2"]["vm"], 0.92617; atol = 1e-3)
         @test isapprox(result["solution"]["bus"]["3"]["vm"], 0.90000; atol = 1e-3)
 
-        @test isapprox(result["solution"]["dcline"]["1"]["p_from"], 10; atol = 1e-3)
-        @test isapprox(result["solution"]["dcline"]["1"]["p_to"], -10; atol = 1e-3)
-
+        @test isapprox(result["solution"]["dcline"]["1"]["pf"],  10; atol = 1e-3)
+        @test isapprox(result["solution"]["dcline"]["1"]["pt"], -10; atol = 1e-3)
+        @test isapprox(result["solution"]["dcline"]["1"]["qf"], -40.3045; atol = 1e-3)
+        @test isapprox(result["solution"]["dcline"]["1"]["qt"],   6.47562; atol = 1e-3)
     end
     @testset "5-bus asymmetric case" begin
         result = run_pf("../test/data/case5_asym.m", ACPPowerModel, ipopt_solver)
@@ -41,7 +24,7 @@
         @test result["status"] == :LocalOptimal
         @test isapprox(result["objective"], 0; atol = 1e-2)
     end
-    @testset "5-bus case with active DC line" begin
+    @testset "5-bus case with hvdc line" begin
         result = run_ac_pf("../test/data/case5_dc.m", ipopt_solver, setting = Dict("output" => Dict("line_flows" => true)))
 
         @test result["status"] == :LocalOptimal
@@ -57,8 +40,8 @@
         @test isapprox(result["solution"]["bus"]["5"]["vm"], 1.0530; atol = 1e-3)
 
 
-        @test isapprox(result["solution"]["dcline"]["1"]["p_from"], 10; atol = 1e-3)
-        @test isapprox(result["solution"]["dcline"]["1"]["p_to"], -8.9; atol = 1e-3)
+        @test isapprox(result["solution"]["dcline"]["1"]["pf"], 10; atol = 1e-3)
+        @test isapprox(result["solution"]["dcline"]["1"]["pt"], -8.9; atol = 1e-3)
 
     end
     @testset "6-bus case" begin
@@ -87,7 +70,7 @@ end
         @test result["status"] == :LocalOptimal
         @test isapprox(result["objective"], 0; atol = 1e-2)
 
-        @test isapprox(result["solution"]["gen"]["1"]["pg"], 144.99; atol = 1e-1)
+        @test isapprox(result["solution"]["gen"]["1"]["pg"], 154.994; atol = 1e-1)
 
         @test isapprox(result["solution"]["bus"]["1"]["va"], 0.00000; atol = 1e-3)
         @test isapprox(result["solution"]["bus"]["2"]["va"], 5.24122; atol = 1e-3)
@@ -120,30 +103,20 @@ end
     @testset "3-bus case" begin
         result = run_pf("../test/data/case3.m", SOCWRPowerModel, ipopt_solver)
 
+        @test result["status"] == :LocalOptimal
+        @test isapprox(result["objective"], 0; atol = 1e-2)
+
         @test result["solution"]["gen"]["1"]["pg"] >= 148.0
+
+        @test isapprox(result["solution"]["gen"]["2"]["pg"], 160.0063; atol = 1e-1)
+        @test isapprox(result["solution"]["gen"]["3"]["pg"], 0; atol = 1e-1)
 
         @test isapprox(result["solution"]["bus"]["1"]["vm"], 1.09999; atol = 1e-3)
         @test isapprox(result["solution"]["bus"]["2"]["vm"], 0.92616; atol = 1e-3)
         @test isapprox(result["solution"]["bus"]["3"]["vm"], 0.89999; atol = 1e-3)
 
-        @test result["status"] == :LocalOptimal
-        @test isapprox(result["objective"], 0; atol = 1e-2)
-    end
-    @testset "3-bus case with active hvdc line" begin
-        result = run_pf("../test/data/case3_dc.m", SOCWRPowerModel, ipopt_solver, setting = Dict("output" => Dict("line_flows" => true)))
-
-        @test result["status"] == :LocalOptimal
-        @test isapprox(result["objective"], 0; atol = 1e-2)
-
-        @test isapprox(result["solution"]["gen"]["2"]["pg"], 160.0063; atol = 1e-1)
-        @test isapprox(result["solution"]["gen"]["3"]["pg"], 0; atol = 1e-1)
-
-        @test isapprox(result["solution"]["bus"]["1"]["vm"], 1.10000; atol = 1e-3)
-        @test isapprox(result["solution"]["bus"]["2"]["vm"], 0.92617; atol = 1e-3)
-        @test isapprox(result["solution"]["bus"]["3"]["vm"], 0.90000; atol = 1e-3)
-
-        @test isapprox(result["solution"]["dcline"]["1"]["p_from"], 10; atol = 1e-3)
-        @test isapprox(result["solution"]["dcline"]["1"]["p_to"], -10; atol = 1e-3)
+        @test isapprox(result["solution"]["dcline"]["1"]["pf"], 10; atol = 1e-3)
+        @test isapprox(result["solution"]["dcline"]["1"]["pt"], -10; atol = 1e-3)
     end
     @testset "5-bus asymmetric case" begin
         result = run_pf("../test/data/case5_asym.m", SOCWRPowerModel, ipopt_solver)